Make OVMS adapter read loaded model size from file
Until this:
https://github.com/kserve/modelmesh-serving/issues/392

is resolved we don't have quick way to work arround the issue that
global file size multiplier is not applicable for different models
and would result in either underutilizing memory or crashes due to
OOM kills.

Having metrics come from OpenVINO would be ideal approach but it won't
be ready in near future.